Project/requirements.txt
streamlit>=1.28.0
torch>=2.0.0
ultralytics>=8.0.0
requests>=2.31.0
Pillow>=9.0.0
numpy>=1.24.0
opencv-python>=4.8.0
pycocotools>=2.0.6
albumentations>=1.3.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
pyyaml>=6.0.0
fastapi==0.104.0
uvicorn==0.29.0
huggingface-hub>=0.25.0
